move classes for consistency
