+++ src/main/scala/edu/cmu/cs/vbc/vbytecode/instructions/Instruction.scala
@@ -3,8 +3,8 @@ package edu.cmu.cs.vbc.vbytecode.instructions
 import edu.cmu.cs.vbc.OpcodePrint
 import edu.cmu.cs.vbc.analysis.VBCFrame
 import edu.cmu.cs.vbc.analysis.VBCFrame.UpdatedFrame
+import edu.cmu.cs.vbc.utils.LiftUtils
 import edu.cmu.cs.vbc.vbytecode._
-import edu.cmu.cs.vbc.vbytecode.util.LiftUtils
 import org.objectweb.asm.MethodVisitor
 import org.objectweb.asm.Opcodes._
